+++ SampleApps/SPMTest/SPMTest/ViewController.swift
@@ -5,14 +5,11 @@ import BraintreeCore
 import BraintreeDataCollector
 import BraintreePaymentFlow
 import BraintreePayPal
+import BraintreeThreeDSecure
 import BraintreeUnionPay
 import BraintreeVenmo
 import PayPalDataCollector
 
 class ViewController: UIViewController {
 
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-    }
 }