+++ Sources/BraintreeThreeDSecure/V2UICustomization/BTThreeDSecureV2UICustomization_Internal.h
@@ -1,0 +1,11 @@
+#if __has_include(<Braintree/BraintreeThreeDSecure.h>)
+#import <Braintree/BTThreeDSecureV2UICustomization.h>
+#else
+#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization.h>
+#endif
+
+@interface BTThreeDSecureV2UICustomization ()
+
+@property (nonatomic, strong) id cardinalValue;
+
+@end