+++ UnitTests/BraintreeThreeDSecureTests/BraintreeThreeDSecureTests-Bridging-Header.h
@@ -6,3 +6,5 @@
 #import <BraintreeThreeDSecure/BTThreeDSecurePostalAddress_Internal.h>
 #import <BraintreeThreeDSecure/BTPaymentFlowDriver+ThreeDSecure_Internal.h>
 #import <BraintreeThreeDSecure/BTThreeDSecureV1BrowserSwitchHelper.h>
+#import <BraintreeThreeDSecure/BTThreeDSecureV2UICustomization_Internal.h>
+#import <BraintreeThreeDSecure/BTThreeDSecureV2BaseCustomization_Internal.h>