+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 # Braintree iOS SDK Release Notes
 
 ## unreleased (5.0.0-beta2)
-* Add SPM support for `BraintreeDataCollector`
+* Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`
 * Bump Kount to v4.0.4.3 pre-release (provides an xcframework for SPM)
 * Bump PPRiskMagnes to v5.10.0 (resolves #564)
 * Fix Xcode 12.3 issue with building PPRiskMagnes.framework for iOS + iOS Simulator
 * Add `paymentTypeCountryCode` to `BTLocalPaymentRequest`
 * Breaking Changes
   * Remove `type` and `nonce` params on `BTApplePayCardNone` initializer
+  * Replace `uiCustomization` with `v2UICustomization` on `BTThreeDSecureRequest`
+  * Introduce new classes for 3DS2 UI customization:
+    * `BTThreeDSecureV2UICustomization`
+    * `BTThreeDSecureV2ButtonCustomization`
+    * `BTThreeDSecureV2LabelCustomization`
+    * `BTThreeDSecureV2TextBoxCustomization`
+    * `BTThreeDSecureV2ToolbarCustomization`
 
 ## 5.0.0-beta1 (2020-12-01)
 * Add support for Swift Package Manager (resolves #462)